+++ HISTORY.rst
@@ -65,6 +65,17 @@
   Unreleased Changes
   ------------------
 
+Crop Production
+===============
+* Both the Percentile and Regression models now issue a warning if any LULC
+  code in the LULC raster is not present in the landcover to crop table, or if
+  any LULC code in the landcover to crop table is not present in the LULC
+  raster (`#925 <https://github.com/natcap/invest/issues/925>`_).
+* The Regression model now correctly validates crop names against the existence
+  of a corresponding regression yield table
+  (`#1723 <https://github.com/natcap/invest/issues/1723>`_).
+
+
 3.15.0 (2025-04-03)
 -------------------
 Highlights